requirements.txt
allure-pytest==2.12.0
assertpy==1.1
Faker==12.3.3
numpy==1.21.5
pandas==1.3.5
polling2==0.5.0
pytest==7.0.1
